Use "files" package.json field instead of npmignore
The npmignore file excludes some items, but is not exhaustive and a lot
of artifacts still get packaged up, leading to a package that currently
takes up to 4.4 MB on disk. You could keep adding entries to npmignore
but it's easy to miss things as the build process evolves.

With this change the on-disk size is down to a little over 200KB. This
includes `/src` plus `CHANGELOG`, `LICENSE`, `package.json`, and
`README` (https://docs.npmjs.com/files/package.json#files).

Fixes https://github.com/auth0/node-auth0/issues/271