[DXEX-455] Allow custom headers to be set in Management/Auth Clients.
In order to better determine what version of the Deploy CLI tool is being used
by our customers, we would like to include the Deploy CLI's version in the
User-agent header as opposed to the current behavior (the version of node under
which the CLI is running).

The Deploy CLI uses ManagementClient to do its work, but the User-agent header
was not configurable before this change. This commit provides a new option
called `headers` that will allow the caller to provide an optional set of
additional headers to outgoing requests.

In the case of the `User-agent` header, if it is not provided by the caller,
the original User-agent string will be used (e.g. `node.js/13.1.2`).